+++ signals/generators/ios/swift/template_methods.py
@@ -1,6 +1,6 @@
 from signals.generators.ios.conversion import get_proper_name
 from signals.generators.ios.ios_template_methods import iOSTemplateMethods
-from signals.generators.ios.swift_generator.swift_parameters import SwiftParameter
+from signals.generators.ios.swift.parameters import SwiftParameter
 from signals.parser.api import GetAPI
