+++ tests/generators/ios/objc/__init__.py
@@ -1,0 +1,1 @@
+__author__ = 'Alex'
