+++ tests/generators/ios/objc/files/EntityMapping.m
