+++ tests/generators/ios/objc/files/GetResponseDescriptor.m
