+++ tests/generators/ios/objc/files/RelationshipMapping.m
