+++ tests/generators/ios/objc/files/SendObject.m
