+++ tests/generators/ios/objc/files/SendObjectWithId.m
