+++ tests/generators/ios/objc/test_template.py
@@ -1,21 +1,16 @@
 from inspect import getmembers, isfunction
 import os
 import unittest
-from datetime import datetime
 from jinja2 import PackageLoader
 from jinja2 import Environment
 from signals.generators.ios.conversion import sanitize_field_name, get_proper_name
-from signals.generators.ios.objc_generator.objc_template import ObjectiveCTemplate
-from signals.generators.ios.objc_generator.objc_template_methods import ObjectiveCTemplateMethods
-from signals.generators.ios.swift_generator.swift_template import SwiftTemplate
-from signals.generators.ios.swift_generator.swift_template_methods import SwiftTemplateMethods
+from signals.generators.ios.objc.template_methods import ObjectiveCTemplateMethods
 from signals.parser.fields import Relationship, Field
 from signals.parser.schema import DataObject, Schema, URL
 from signals.parser.api import GetAPI, PatchAPI, PostAPI
 from tests.utils import create_dynamic_schema
 
 
-# TODO: Write tests for Swift Templates
 class TemplateTestCase(unittest.TestCase):
     def setUp(self):
         self.jinja2_environment = Environment(loader=PackageLoader('signals.generators.ios', 'templates/objc/'),
@@ -243,70 +238,3 @@ class TemplateTestCase(unittest.TestCase):
             'data_object': post_object,
             'relationship': relationship
         })
-
-    def test_objective_c_data_model_header_template(self):
-        schema = Schema("./tests/files/test_schema.json")
-        self.assertTemplateEqual('data_model.h.j2', 'DataModel.h', {
-            'schema': schema,
-            'VIDEO_FIELD': Field.VIDEO,
-            'IMAGE_FIELD': Field.IMAGE,
-            'today': datetime.today(),
-            'endpoints': URL.URL_ENDPOINTS.keys(),
-        }, expected_context=(
-            datetime.today().strftime('%m/%d/%Y'),
-        ))
-
-    def test_objective_c_data_model_implementation_template(self):
-        schema = Schema("./tests/files/test_schema.json")
-        self.assertTemplateEqual('data_model.m.j2', 'DataModel.m', {
-            'project_name': "TestProject",
-            'schema': schema,
-            'VIDEO_FIELD': Field.VIDEO,
-            'IMAGE_FIELD': Field.IMAGE,
-            'today': datetime.today(),
-            'endpoints': URL.URL_ENDPOINTS.keys(),
-            'request_objects': ObjectiveCTemplate.get_request_objects(schema.data_objects),
-            'sanitize_field_name': sanitize_field_name
-        }, expected_context=(
-            datetime.today().strftime('%m/%d/%Y'),
-        ))
-
-
-class SwiftTemplateTestCase(unittest.TestCase):
-    def setUp(self):
-        self.jinja2_environment = Environment(loader=PackageLoader('signals.generators.ios', 'templates/swift/'),
-                                              extensions=['jinja2.ext.with_'],
-                                              trim_blocks=True,
-                                              lstrip_blocks=True)
-
-    def assertTemplateEqual(self, template_name, expected_template, context, expected_context=None):
-        current_directory = os.path.dirname(__file__)
-        expected_file_path = os.path.join(current_directory, "files", expected_template)
-        with open(expected_file_path, "r") as expected_template_file:
-            template = self.jinja2_environment.get_template(template_name)
-            # Registers all methods in template_methods.py with jinja2 for use
-            context.update({name: method for name, method in getmembers(SwiftTemplateMethods, isfunction)})
-            context.update({'get_proper_name': get_proper_name,
-                            'sanitize_field_name': sanitize_field_name
-                            })
-            template_output = template.render(**context)
-            expected_template_out = expected_template_file.read()
-            if expected_context:
-                # Using %s formatting syntax, since format uses {'s and the outputted code contains many
-                expected_template_out = expected_template_out % expected_context
-            self.assertEqual(template_output, expected_template_out)
-
-    def test_swift_data_model_implementation_template(self):
-        schema = Schema("./tests/files/test_schema.json")
-        self.assertTemplateEqual('data_model.swift.j2', 'DataModel.swift', {
-            'project_name': "TestProject",
-            'schema': schema,
-            'VIDEO_FIELD': Field.VIDEO,
-            'IMAGE_FIELD': Field.IMAGE,
-            'today': datetime.today(),
-            'endpoints': URL.URL_ENDPOINTS.keys(),
-            'request_objects': SwiftTemplate.get_request_objects(schema.data_objects),
-            'sanitize_field_name': sanitize_field_name
-        }, expected_context=(
-            datetime.today().strftime('%m/%d/%Y'),
-        ))
