+++ tests/generators/ios/swift/files/DataModel.swift
